Move shareable Google API client code from google.py to new google_base.py. Add europe-west4-a to available zones.
PiperOrigin-RevId: 189089372
